+++ src/design/plone/contenttypes/events/common.py
@@ -120,6 +120,11 @@ SUBFOLDERS_MAPPING = {
         },
         {"id": "altre-cariche", "title": "Altre cariche", "allowed_types": ("File",)},
         {"id": "incarichi", "title": "Incarichi", "allowed_types": ("Incarico",)},
+        {
+            "id": "altri-documenti",
+            "title": "Altri documenti",
+            "allowed_types": ("File", "Image", "Link"),
+        },
     ],
     "Pratica": [
         {