+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 6.2.9 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Aggiunto la cartella "Altri Documenti" sulla persona
+  [lucabel]
 
 
 6.2.8 (2024-04-22)